+++ backend/dependencies.py
@@ -1,0 +1,24 @@
+from typing import Generator
+
+from passlib.context import CryptContext
+from sqlalchemy.orm import Session
+
+from .database import SessionLocal
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+
+def get_db() -> Generator[Session, None, None]:
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+def get_password_hash(password: str) -> str:
+    return pwd_context.hash(password)
+
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    return pwd_context.verify(plain_password, hashed_password)
