+++ backend/routers/assets.py
@@ -1,0 +1,53 @@
+from typing import List
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+
+from .. import models, schemas
+from ..dependencies import get_db
+
+router = APIRouter(prefix="/assets", tags=["assets"])
+
+
+@router.post("/", response_model=schemas.AssetRead)
+def create_asset(asset: schemas.AssetCreate, db: Session = Depends(get_db)):
+    db_asset = models.Asset(name=asset.name, owner=asset.owner)
+    db.add(db_asset)
+    db.commit()
+    db.refresh(db_asset)
+    return db_asset
+
+
+@router.get("/", response_model=List[schemas.AssetRead])
+def list_assets(db: Session = Depends(get_db)):
+    return db.query(models.Asset).all()
+
+
+@router.get("/{asset_id}", response_model=schemas.AssetRead)
+def get_asset(asset_id: int, db: Session = Depends(get_db)):
+    asset = db.query(models.Asset).filter(models.Asset.id == asset_id).first()
+    if not asset:
+        raise HTTPException(status_code=404, detail="Asset not found")
+    return asset
+
+
+@router.put("/{asset_id}", response_model=schemas.AssetRead)
+def update_asset(asset_id: int, asset_update: schemas.AssetUpdate, db: Session = Depends(get_db)):
+    asset = db.query(models.Asset).filter(models.Asset.id == asset_id).first()
+    if not asset:
+        raise HTTPException(status_code=404, detail="Asset not found")
+    asset.name = asset_update.name
+    asset.owner = asset_update.owner
+    db.commit()
+    db.refresh(asset)
+    return asset
+
+
+@router.delete("/{asset_id}")
+def delete_asset(asset_id: int, db: Session = Depends(get_db)):
+    asset = db.query(models.Asset).filter(models.Asset.id == asset_id).first()
+    if not asset:
+        raise HTTPException(status_code=404, detail="Asset not found")
+    db.delete(asset)
+    db.commit()
+    return {"detail": "deleted"}
